Highlight allowed moves and move pawn
Created a system where once a chesspiece is selected the squares it is allowed to move to is highlighted. So the player gets a visual feedback of what moves he can make. Also added the basic allowed move for pawn except for promotion and en passant. There is also the system where one piece can capture the opponents piece if it gets in the way.