Check system added
So now it's possible to check the opponent. Once an oppenent is checked he can not make a move unless that move unchecks him. There is still no checkmate system but the oppenent is blocked if that is the case. Also fixed the enpassant. Now the same team can not enpassant themselves.