Modified the check system added game win condition
The previous check system was really inefficient. So made a new data class called Move which contains all the possible move of a particular side. If any move results in his own king getting checked then that move is not legal. If any side has zero possible moves than the opposite team has won. Also added the castling move with necessary conditions. Castling can't be done if in check or if there are other pieces in between king and rook or if king has moved. The only thing left is the game draw condition. Then I'll be good to start working on the AI.