Fixed the draw condition. Started working on a basic AI
So the draw system is fixed now. Every time a side runs out of moves it is checked wheather that side is in check. If not the match is draw. Also started the AI. It's obviously terrible at this point. But it can calculate which move will increase new moves and decrease moves for the opponent. Hope to improve on it.